requirements.txt
boto3==1.26.137
botocore==1.29.137
jmespath==1.0.1
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
urllib3==1.26.15
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
billiard==3.6.4.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cron-descriptor==1.2.35
Django==4.1.7
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-timezone-field==5.0
gunicorn==20.1.0
kombu==5.2.4
prompt-toolkit==3.0.38
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
rabbitmq==0.2.0
redbird==0.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0